baseProxy API change + cleanup
BaseProxy now has inner class holding status information.
This class is intended to logically separate state information from
other variables, it's clearer to read and will be extended in the future
with state machine variables.

API change is quite significant and this patch has the same description
in socle and smithproxy projects.