fix memory pool exhaustion bug
Problem:
Certain smaller pools are allocated based on 256B pool size.
The calculation is done on startup, and could pool sizes could be potentially extended
in runtime.
When pointer is being released, it must be returned to the correct pool, based on
its allocation in bytes. However function doing the pool selection was calculating
with wrong data, and didn't select correct pool, therefore returning nullptr.
As a result, chunk couldn't have been released and was freed as the last resort action.

Solution:
Simply return the right pool and don't be too smart with calculating pool size.
On the other hand if MEMPOOL_DEBUG is defined, throw exception if we miss the pool.