fix cpu spikes when right connection waits for left's ClientHello peek
Problem: on SSL connect() call, we read peer's ClientHello by peeking his socket.
We expect it to already be here, but in case it's not, we enter busy loop.
Since we can't read connect() socket data yet, and there are data pending,
therefore socket with EPOLLIN flag.

Solution: remove connect() socket from monitored, and rather set it for
rescan. Put it back to monitored socket once ClientHello is received and processed.

Signed-off-by: astib <astib@mag0.net>