fix TCP connection vs state race
This was a long-standing problem with symptom of prematurely closed
sessions.

TL;DR:

- fix redundant TCP state `getsockopt()` calls, fix INPROGRESS behaviour.

- fix `accept()` handling of EAGAIN

Problem was however not trivial:

On regular internet speeds this didn't manifest often, the same applied
to slower VM testbeds. That being said only after moving to my new ISP
with optics-like latencies, I figured out.

We were asking on several places for TCP status.

Mistake 1/ was to ask each time, even though we already knew the status checked before.
Now, the state could migrate to EINPROGRESS again due to external socket events (data arrived, etc).
So asking consecutively can result in various handshake delays in upper layers, like TLS.

Mistake 2/ EINPROGRESS is not a error, BUT we should immediately
check the status of socket by querying it with epoll. We luckily don't have to do that,
because we have it already in epoller.in_ or out_ set.
Reporting connected state by checking epoller sets makes thing a lot faster.

Mistake 3/ EAGAIN in `accept()` is not an error, but understanding it as a positive socket
operation is a problem. New socket fd -1 can be hardly considered as an desired outcome.

Signed-off-by: astib <astib@mag0.net>